backend/requirements.txt
# AI Gatekeeper Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Google Cloud & Vertex AI
google-cloud-aiplatform==1.42.0
google-cloud-storage==2.14.0
vertexai==0.0.1

# ElevenLabs
elevenlabs==0.2.27
websockets==12.0

# Twilio
twilio==8.11.0

# Database (Supabase)
supabase==2.3.1
postgrest==0.13.0

# Async & Data
httpx==0.24.1
aiohttp==3.9.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Audio Processing
pyaudio==0.2.14
# audioop is built-in, do not list here

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging & Monitoring
structlog==24.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.23.2
black==23.12.1
ruff==0.1.9
